requirements.txt
solana==0.36.5
solders==0.25.0
requests==2.32.3
websockets==14.2
python-dotenv==1.0.1
loguru==0.7.3
aiohttp==3.11.12
base58==2.1.1
